+++ src/components/Common/Badges.js
@@ -1,45 +1,28 @@
-import React, { useContext } from "react";
+import React from "react";
 import Image from "react-bootstrap/Image";
 import { Link } from "react-router-dom";
 
 import Medal from "./images/medal.svg";
 import Star from "./images/star.svg";
-import Logout from "./images/logout.svg";
-import { UserContext } from "../../contexts/UserContext";
 
 const styles = {
 	badgeContainer: {
-		display: "flex"
+		display: "flex",
 	},
 	badgeItem: {
 		color: "white",
-		margin: "0 10px"
+		margin: "0 10px",
 	},
 	badgeIcon: {
 		width: "35px",
-		margin: "0 10px"
+		margin: "0 10px",
 	},
 	link: {
-		textDecoration: "none"
-	}
+		textDecoration: "none",
+	},
 };
 
-export default function(props) {
-	const student = useContext(UserContext);
-	const clearSess = () => {
-		const s = {
-			matric: undefined,
-			name: undefined,
-			class: undefined,
-			current_progress: undefined,
-			avatar_url: undefined,
-			stars: undefined,
-			medals: undefined
-		};
-		student.setStudent(s);
-		window.alert("You've been succefully logged out!");
-	};
-
+export default function (props) {
 	return (
 		<div style={styles.badgeContainer}>
 			<Link key="leader" to={`/leader`} style={styles.link}>
@@ -50,11 +33,6 @@ export default function(props) {
 				<span style={styles.badgeItem}>
 					<Image src={Star} style={styles.badgeIcon} draggable={false} />
 					{props.stars}
-				</span>
-			</Link>
-			<Link to={`/login`} style={styles.link} onClick={clearSess}>
-				<span>
-					<Image src={Logout} style={styles.badgeIcon} draggable={false} />
 				</span>
 			</Link>
 		</div>
