+++ src/components/Navbar/NavbarContainer.js
@@ -1,14 +1,36 @@
 import React, { useContext } from "react";
 import { Link } from "react-router-dom";
 import { Container, Navbar } from "react-bootstrap";
-import { Music } from "../Common";
+import { Music, Badges } from "../Common";
+import Logout from "./images/logout.svg";
 
 import { UserContext } from "../../contexts/UserContext";
-import { Badges } from "../Common";
+
+const styles = {
+	logoutSpan: {
+		color: "white",
+		cursor: "pointer",
+	},
+	logoutIcon: { width: "35px", marginRight: "5px" },
+};
 
 export default function () {
 	const { student } = useContext(UserContext);
 	const { matric, medals = "u", stars = "u" } = student;
+
+	const clearSess = () => {
+		const s = {
+			matric: undefined,
+			name: undefined,
+			class: undefined,
+			current_progress: undefined,
+			avatar_url: undefined,
+			stars: undefined,
+			medals: undefined,
+		};
+		student.setStudent(s);
+		alert("You've been succefully logged out!");
+	};
 
 	return (
 		<Navbar expand="lg" variant="dark" bg="dark" fixed="top">
@@ -18,6 +40,10 @@ export default function () {
 				</Navbar.Brand>
 				<Music />
 				<Badges medals={medals} stars={stars} />
+				<span style={styles.logoutSpan} onClick={clearSess}>
+					<img src={Logout} style={styles.logoutIcon} draggable={false} />
+					Logout
+				</span>
 			</Container>
 		</Navbar>
 	);
